+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Python class for reading and writing
 
 - `NLLGrid.recenter()` method to move origin of grid cartesian system to the
   grid center
+- `NLLGrid.rotate()` method to rotate the grid horizontally around its center
 - Added `TRANS_MERC` and `AZIMUTHAL_EQUIDIST` projections
 - Fix projections when `map_rot != 0`